update ball and bars collision checking

+++ app/src/main/java/com/example/charlie/g1_roll_it_in/gameModel/GameObject.java
@@ -53,7 +53,7 @@ public abstract class GameObject {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
